Updated to use pymongo
To install mongo properly (so it *should* work -> ```If you are on ubuntu, try this. If you are on Windows, try this (Keep this console separate and open). If you are on another Linux distro, try this. You will also need to install the python module by doing pip3 install pymongo.```

- Further testing needed before pushing to live