Pswdchannel fix (#37)
* Updated to use pymongo

To install mongo properly (so it *should* work -> ```If you are on ubuntu, try this. If you are on Windows, try this (Keep this console separate and open). If you are on another Linux distro, try this. You will also need to install the python module by doing pip3 install pymongo.```

* Added Feedback on pswd removal

Upon pswd removal, a msg is sent confirming the removal of a pswd

* Update info.json

Added more info to it. (making it more clear about the cpu usage)
Added pymongo requirement


* Removed the c# comment syntax 

Removed the line commented in c# syntax 
Line wasn't needed as it would result in logical bug of sending a message when the role was removed.